More game modes
added reverse and mixed game mode
