+++ package.json
@@ -14,11 +14,13 @@
     "react": "18.2.0",
     "react-confetti": "6.2.2",
     "react-dom": "18.2.0",
+    "react-router-dom": "^7.2.0",
     "react-scripts": "^5.0.1"
   },
   "devDependencies": {
     "@types/react": "18.2.38",
     "@types/react-dom": "18.2.15",
+    "@types/react-router-dom": "^5.3.3",
     "typescript": "4.4.4"
   },
   "scripts": {
