+++ src/App.tsx
@@ -1,10 +1,17 @@
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import GameModeSelector from './GameModeSelector';
 import HiraganaCard from "./HiraganaCard";
-//import HiraganaCard2 from "./HiraganaCard2";
+import Game from './Game';
 
 export default function App() {
   return (
-    <div className="App">
-      <HiraganaCard />
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<GameModeSelector />} />
+        <Route path="/hiragana-quiz" element={<Game mode="hiragana" />} />
+        <Route path="/reverse-mode" element={<Game mode="romaji" />} />
+        <Route path="/mixed-quiz" element={<Game mode="mixed" />} />
+      </Routes>
+    </Router>
   );
 }
