+++ src/HiraganaCard.css
@@ -9,10 +9,10 @@
 }
 
 .hiragana-card {
-  width: 300px;
-  height: 400px;
+  width: 100%;
+  max-width: 500px;
+  margin: 0 auto;
   perspective: 1000px;
-  position: relative;
 }
 
 .card-content {
@@ -23,7 +23,7 @@
   box-shadow: 0 8px 30px rgba(0, 0, 0, 0.2);
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
+  justify-content: center;
   padding: 30px;
   position: relative;
   backdrop-filter: blur(5px);
@@ -43,6 +43,7 @@
 }
 
 .hiragana {
+  height: 200px;
   font-size: 120px;
   color: #333;
   text-align: center;
@@ -56,7 +57,7 @@
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   gap: 10px;
-  margin-top: 20px;
+  margin-top: 0px;
   margin-bottom: 20px;
 }
 
@@ -157,4 +158,62 @@
   100% {
     transform: rotateY(0);
   }
+}
+
+.hiragana-title {
+  font-size: 2.5rem;
+  color: #333;
+  text-align: center;
+  margin: 2rem 0;
+  font-weight: 700;
+}
+
+.mode-grid {
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+  padding: 20px;
+  align-items: center;
+  min-height: 60vh;
+}
+
+.mode-card {
+  background: rgba(255, 255, 255, 0.98);
+  border-radius: 15px;
+  padding: 2rem;
+  margin: 1rem;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
+}
+
+.mode-card:hover {
+  transform: translateY(-5px);
+  box-shadow: 0 8px 25px rgba(0, 0, 0, 0.15);
+}
+
+.mode-emoji {
+  font-size: 3rem;
+  margin-bottom: 1rem;
+}
+
+.mode-title {
+  font-size: 1.5rem;
+  color: #2d3748;
+  margin-bottom: 0.75rem;
+  font-weight: 600;
+}
+
+.mode-description {
+  color: #4a5568;
+  text-align: center;
+  white-space: pre-line;
+  font-size: 2.5rem;
+  line-height: 1.5;
+}
+
+.hover-scale {
+  transition: transform 0.2s ease-in-out;
+}
+
+.hover-scale:hover {
+  transform: scale(1.02);
 }
